requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.15.4
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==4.9.3
numpy==2.2.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parsel==1.9.1
pefile==2023.2.7
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
service-identity==24.1.0
setuptools==71.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tldextract==5.1.2
Twisted==24.3.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.4
uvicorn==0.35.0
w3lib==2.2.1
watchfiles==1.1.0
websockets==15.0.1
zope.interface==6.4.post2
fastapi
sqlalchemy
python-jose[cryptography] 
python-multipart 
aiofiles
itsdangerous
Jinja2
asyncpg
psycopg2-binary
passlib
httpx
argon2-cffi
pytest
pytest-asyncio
pytest-env
pytest-async-sqlalchemy
aiosqlite